moved logic to selector.
